Cleanup scripts

+++ plugins/owncloud/scripts/pre-install.sh
@@ -4,4 +4,6 @@
 owncloud_pbi_path=/usr/pbi/owncloud-$(uname -m)
 
 # Backup ownCloud config
-mv ${owncloud_pbi_path}/www/owncloud/config/config.php /media
+if [ -f "${owncloud_pbi_path}/www/owncloud/config/config.php" ]; then
+	mv ${owncloud_pbi_path}/www/owncloud/config/config.php /media
+fi
